backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pypdf2>=3.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain>=0.0.267
torch>=2.0.0
spacy>=3.5.0
accelerate>=0.20.0
bitsandbytes>=0.40.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
psutil>=5.9.0  # For memory tracking